feat: add metadata for metrics
