+++ Cargo.toml
@@ -50,6 +50,7 @@ async-trait = "0.1"
 async_once = "0.2"
 awc = "3.1"
 base64 = "0.21"
+blake3 = "1.3"
 bytes = "1.4"
 chrono = { version = "0.4", default-features = false, features = ["clock"] }
 clap = { version = "4.1", default-features = false, features = [
@@ -71,6 +72,7 @@ flatten-json-object = "0.6"
 futures = "0.3"
 get_if_addrs = "0.5"
 glob = "0.3"
+hex = "0.4"
 http-auth-basic = "0.3"
 lazy_static = "1.4"
 log = "0.4"
